Move stats into base module
